backend/requirements.txt
pydantic==1.10.12
fastapi==0.103.2
uvicorn==0.23.2
pymongo==4.15.4
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.1
openpyxl==3.1.2
starlette==0.27.0
typing-extensions==4.12.0
cloudinary
Pillow==10.0.1
numpy==1.24.3
requests==2.31.0
deepface==0.0.77
tensorflow==2.13.0
opencv-python==4.8.0.76
pydantic-settings
